Refactor: Decouple main.py monolith into scalable modular architecture
