+++ backend/core/config.py
@@ -1,0 +1,54 @@
+import os
+from typing import Any
+
+RESPONSE_TIME_HEADER = "X-Response-Time-ms"
+DEFAULT_SLOW_RESPONSE_THRESHOLD_MS = 1000.0
+CACHE_TTL_SECONDS = 300
+CACHE_CONTROL_VALUE = f"public, max-age={CACHE_TTL_SECONDS}"
+MAX_UPLOAD_BYTES = int(os.environ.get("MAX_UPLOAD_BYTES", str(5 * 1024 * 1024)))
+MAX_SEARCH_QUERY_LENGTH = 120
+CACHE_MAX_ENTRIES = int(os.environ.get("CACHE_MAX_ENTRIES", "2000"))
+ADMIN_API_TOKEN_ENV = "ADMIN_API_TOKEN"
+
+MOCK_PRODUCTS = [
+    {
+        "id": 1,
+        "title": "Acoustic Noise-Cancelling Headphones",
+        "description": "Premium over-ear headphones with active noise cancellation.",
+        "category": "Electronics",
+        "rating": 4.8,
+        "avg_sentiment": 0.85,
+        "review_count": 245,
+        "price": 1299,
+    },
+    {
+        "id": 2,
+        "title": "Ergonomic Mechanical Keyboard",
+        "description": "Tactile switches, RGB backlighting, and a comfortable wrist rest.",
+        "category": "Electronics",
+        "rating": 4.5,
+        "avg_sentiment": 0.65,
+        "review_count": 189,
+        "price": 799,
+    },
+    {
+        "id": 3,
+        "title": "Portable Fitness Tracker",
+        "description": "Track heart rate, sleep, and workouts from your wrist.",
+        "category": "Health",
+        "rating": 4.2,
+        "avg_sentiment": 0.42,
+        "review_count": 128,
+        "price": 499,
+    },
+]
+
+
+def _get_slow_response_threshold_ms() -> float:
+    try:
+        return float(os.environ.get("RESPONSE_TIME_SLOW_MS", DEFAULT_SLOW_RESPONSE_THRESHOLD_MS))
+    except ValueError:
+        return DEFAULT_SLOW_RESPONSE_THRESHOLD_MS
+
+def _cache_key(*parts: Any) -> str:
+    return ":".join(str(part).strip().lower() for part in parts)
